refactor: create srpcLib module and ajust imporets
