+++ srpcLib/stub_generator/srpc_server_stub_gen.py
@@ -1,6 +1,6 @@
 import logging
 
-from stub_generator.srpc_stub_utils import DEFAULT_BINDER_PORT
+from srpc_stub_utils import DEFAULT_BINDER_PORT
 
 logger = logging.getLogger(__name__)
 log_path = "./srpc_server_metrics.log"
